+++ examples/f90/issue_2449_openacc_fixed.f
@@ -1,8 +1,9 @@
 C Fixed-form OpenACC directive test
       program testopenacc
+      use iso_fortran_env, only: dp => real64
       implicit none
       integer :: i, n
-      real :: a(100), b(100), c(100)
+      real(dp) :: a(100), b(100), c(100)
 
       n = 100
 !$acc parallel loop
